backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
psycopg2-binary==2.9.11
python-decouple==3.8
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.5.1
gunicorn==22.0.0
whitenoise==6.6.0
django-filter==24.1
